+++ HLSpriteKit/HLSpriteKit.h
@@ -36,6 +36,7 @@
 #import "HLTiledNode.h"
 #import "HLToolbarNode.h"
 #import "HLUglyShuffler.h"
+#import "HLWrapLayoutManager.h"
 #import "SKLabelNode+HLLabelNodeAdditions.h"
 #import "SKNode+HLAction.h"
 #import "SKNode+HLLayoutManager.h"